+++ src/main/java/cs309/repo/ConnectionRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by jeffrey on 1/31/16.
  */
-public interface ConnectionRepository extends JpaRepository<Connection,Long> {
+public interface ConnectionRepository extends JpaRepository<Connection,Integer> {
 }
